### 2024-02-06 #6327 @danielskatz
In the submission note from the author:

>The algorithm behind the paper has recently been published in "Numerical evaluation of oscillatory integrals via automated steepest descent contour deformation", Journal of Computational Physics 501, 2024. That paper contained numerical experiments using an earlier version of PathFinder, and we gave roughly ten lines of sample code to highlight the versatility and robustness of the algorithm. However, that paper was primarily about the algorithm, not the software. When that paper was written, the code lacked detailed documentation, examples, unit testing, and a stable MEX compilation process. Moreover, it was not Octave compatible, so (in a sense) not open source. These issues have since been addressed, and I believe the software is now worthy of consideration for publication in JOSS.

This is https://doi.org/10.1016/j.jcp.2024.112787 with a preprint at https://arxiv.org/abs/2307.07261
