### 2024-02-06 #6327 @danielskatz
👋 @AndrewGibbs - thanks for your submission. Before we proceed, I want the JOSS editorial board to weigh in on the previous publication.  You should hear back in a week or two.
