### 2024-02-27 #6327 @danielskatz
Thanks @AndrewGibbs - I'll bring this back to the editors and ask them to take another look at it.

I also apologize for my overly brief message. I was just returning from vacation and trying to deal with too many things at once. I should have also said

> Our decision does not mean that your software is not useful to the research community. Please see https://joss.readthedocs.io/en/latest/submitting.html#other-venues-for-reviewing-and-publishing-software-packages for other suggestions for how you might receive credit for your work.
