### 2024-03-01 #6327 @danielskatz
@AndrewGibbs - We have reconsidered, and in particular, discussed the issue of duplicate submission in some detail. In this case, since your paper was an algorithm paper and not a software paper, which was not full appreciated by all of us initially, we have decided to accept this as in scope for review, and I will proceed with this process as soon as possible.
