### 2024-04-13 e35e7ce
Added other OSs to CI.yml
