+++ .github/workflows/CI.yml
@@ -3,7 +3,10 @@ on: [push]
 jobs:
   my-job:
     name: Run MATLAB Tests and Generate Artifacts
-    runs-on: ubuntu-latest       
+    runs-on: 
+      - ubuntu-latest
+      - macOS-latest
+      - windows-latest       
     steps:
       - name: Check out repository
         uses: actions/checkout@v3
