### 2024-04-13 b273fc8
Update CI.yml matrix
