+++ .github/workflows/CI.yml
@@ -3,10 +3,10 @@ on: [push]
 jobs:
   my-job:
     name: Run MATLAB Tests and Generate Artifacts
-    runs-on: 
-      - ubuntu-latest
-      - macOS-latest
-      - windows-latest       
+    strategy:
+      matrix:
+        os: [ubuntu-latest, windows-latest, macos-latest]
+    runs-on: ${{ matrix.os }}   
     steps:
       - name: Check out repository
         uses: actions/checkout@v3
