### 2024-04-17 d8638e6
added MinGW-w64 to CI
