### 2024-04-17 403a876
added win compiler to CI.yml
