+++ .github/workflows/CI.yml
@@ -12,6 +12,8 @@ jobs:
         uses: actions/checkout@v3
       - name: Set up MATLAB
         uses: matlab-actions/setup-matlab@v1
+        with:
+          products: MATLAB_Support_for_MinGW-w64_C/C++/Fortran_Compiler 
       - name: Run tests and generate artifacts
         uses: matlab-actions/run-tests@v1
         with:
