+++ .github/workflows/CI.yml
@@ -5,15 +5,13 @@ jobs:
     name: Run MATLAB Tests and Generate Artifacts
     strategy:
       matrix:
-        os: [ubuntu-latest, windows-latest, macos-latest]
+        os: [macos-latest]
     runs-on: ${{ matrix.os }}   
     steps:
       - name: Check out repository
         uses: actions/checkout@v3
       - name: Set up MATLAB
         uses: matlab-actions/setup-matlab@v1
-        with:
-          products: MATLAB_Support_for_MinGW-w64_C/C++/Fortran_Compiler 
       - name: Run tests and generate artifacts
         uses: matlab-actions/run-tests@v1
         with:
