### 2024-04-17 758a1a1
Update CI.yml
Changed @v1 to @v2
