+++ .github/workflows/CI.yml
@@ -11,7 +11,7 @@ jobs:
       - name: Check out repository
         uses: actions/checkout@v3
       - name: Set up MATLAB
-        uses: matlab-actions/setup-matlab@v1
+        uses: matlab-actions/setup-matlab@v2
         with:
           products: MATLAB_Support_for_MinGW-w64_C/C++/Fortran_Compiler 
       - name: Run tests and generate artifacts
