### 2024-04-17 0a64348
Restricted CI.yml to only windows OS
