+++ .github/workflows/CI.yml
@@ -5,7 +5,7 @@ jobs:
     name: Run MATLAB Tests and Generate Artifacts
     strategy:
       matrix:
-        os: [ubuntu-latest, windows-latest, macos-latest]
+        os: [windows-latest]
     runs-on: ${{ matrix.os }}   
     steps:
       - name: Check out repository
