### 2024-04-17 fd851dd
Update Test_all.m
Oops - used wrong option last time
