+++ test/Test_all.m
@@ -1,8 +1,8 @@
 % check correct compiler is being used in CI log
 mex -setup C;
 if ispc
-  mingw64XmlPath = fullfile(matlabroot, 'bin', 'win64', 'mexopts', 'mingw64.xml');
-  mex -setup:D:\hostedtoolcache\windows\MATLAB\2024.1.999\x64\bin\win64\mexopts\msvc2022.xml C;
+  % mingw64XmlPath = fullfile(matlabroot, 'bin', 'win64', 'mexopts', 'mingw64.xml');
+  mex -setup:D:\hostedtoolcache\windows\MATLAB\2024.1.999\x64\bin\win64\mexopts\mingw64.xml C
   % mex('-setup', mingw64XmlPath, 'C');
   mex -setup C;
 end
