### 2024-04-18 df84e71
fixed other_stationary_points_len warning
