+++ src/mex_compilation/sd_path_ode_euler.c
@@ -56,7 +56,7 @@ void mexFunction(int nlhs, mxArray *plhs[],
     double complex otherSPs[max_1_or_other_SPs_len];
     double otherSPs_radii[other_stationary_points_len];
 
-    if (otherSPs == 0)
+    if (other_stationary_points_len == 0)
     {
         // in the case where there are no other SPs, add fictitious SP at infinity,
         // to keep things simple later on.
