### 2024-04-18 d189aa6
Update rootfinding_notes.md
Corrected md so that it renders in GitHub
