### 2024-04-18 e726116
added MinGW-64 instructions to readme
