### 2024-04-18 387c62c
added MinGW-64 instructions to readme
