+++ .github/workflows/CI.yml
@@ -12,6 +12,7 @@ jobs:
         uses: actions/checkout@v3
       - name: Set up MATLAB
         uses: matlab-actions/setup-matlab@v2
+        if: runner.os == 'Windows'
         with:
           products: MATLAB_Support_for_MinGW-w64_C/C++/Fortran_Compiler 
       - name: Run tests and generate artifacts
