### 2024-04-18 2e2e63b
added auto compiler switch option, for CI
