### 2024-04-18 7e2c9c6
added auto compiler switch option, for CI
