### 2024-04-18 aedc91a
adding paths before compiler switch
