+++ test/Test_all.m
@@ -1,5 +1,7 @@
 % check correct compiler is being used in CI log
 mex -setup C;
+% add pathfinder files to path
+addpath('..');
 if ispc
   % mingw64XmlPath = fullfile(matlabroot, 'bin', 'win64', 'mexopts', 'mingw64.xml');
 %   mex -setup:D:\hostedtoolcache\windows\MATLAB\2024.1.999\x64\bin\win64\mexopts\mingw64.xml C
@@ -7,9 +9,6 @@ if ispc
   % mex('-setup', mingw64XmlPath, 'C');
   mex -setup C;
 end
-
-% add pathfinder files to path
-addpath('..');
 addPaths;
 compileAll;
 show_text = false;
