+++ .github/workflows/CI.yml
@@ -5,13 +5,15 @@ jobs:
     name: Run MATLAB Tests and Generate Artifacts
     strategy:
       matrix:
-        os: [macos-13]
+        os: [macos-12]
     runs-on: ${{ matrix.os }}   
     steps:
       - name: Check out repository
         uses: actions/checkout@v3
       - name: Set up MATLAB
-        uses: matlab-actions/setup-matlab@v1
+        uses: matlab-actions/setup-matlab@v2
+        with:
+          release: R2022a
       - name: Run tests and generate artifacts
         uses: matlab-actions/run-tests@v1
         with:
