### 2024-04-18 53dd82b
adding paths before compiler switch pt2
