### 2024-04-18 e4c7d25
outputting compiler names, debugging via CI...
