### 2024-04-18 89d0204
outputting compiler names, debugging via CI pt2
