### 2024-04-18 763e4c6
Update switchWindowsCompiler.m
outputting number of compilers in CI
