+++ src/mex_compilation/switchWindowsCompiler.m
@@ -4,8 +4,8 @@ function switchWindowsCompiler()
     
     % Find the MinGW-w64 compiler
     mingw_compiler = [];
-    disp(numel(compilers));
-    for i = 1:numel(compilers)
+    disp(length(compilers));
+    for i = 1:length(compilers)
         disp(compilers(i).Name);
         if contains(compilers(i).Name, 'MinGW-w64')
             mingw_compiler = compilers(i);
