### 2024-04-18 70ca001
new compiler switch strategy using eval etc
