### 2024-04-18 3d2eb0a
new compiler switch strategy using eval etc pt2
