### 2024-04-18 808b40a
trying continue-on-error instead
