### 2024-04-18 d912676
attempt to fix CI using reg exprs
