### 2024-04-18 a0bd58f
attempt to fix CI using reg exprs pt 2
