### 2024-04-18 0ca8e19
finally appear to have auto compiler switching working
