### 2024-04-18 97c5470
finally appear to have auto compiler switching working
