### 2024-04-19 20a5dde
added new test to check latest windows bug
