### 2024-05-22 #6327 @olexandr-konovalov
Hello @kwwette and @@ketch - would one or both of you be able to review this submission for JOSS, please?

If not, could you possibly recommend someone whom we may approach regarding this?
