### 2024-05-22 #6327 @olexandr-konovalov
Hi @YehorYudinIPP - would you be able to review this submission for JOSS, please? If not, could you possibly recommend someone whom we may approach regarding this?
