### 2024-05-23 #6327 @YehorYudinIPP
> Hi @YehorYudinIPP - would you be able to review this submission for JOSS, please? If not, could you possibly recommend someone whom we may approach regarding this?

I have some time on my hands for that, but I have never reviewed a software paper before.
