### 2024-05-23 #6327 @kwwette
> Hello @kwwette and @@ketch - would one or both of you be able to review this submission for JOSS, please?
>
> If not, could you possibly recommend someone whom we may approach regarding this?

Unfortunately I'm very busy for the next month or so. If you're still looking for a 2nd reviewer in July, ping me again and I might be able to help.
