### 2024-05-24 #6327 @YehorYudinIPP
> @YehorYudinIPP thanks for the reply - the review takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. JOSS reviews involve downloading and installing the software, and inspecting the repository and submitted paper for key elements (see https://joss.readthedocs.io/en/latest/review_criteria.html for further details). Editors and reviewers post comments on the Review issue, and authors respond to the comments and improve their submission until acceptance (or withdrawal, if they feel unable to satisfy the review). You can find more information in the reviewer's guide at https://joss.readthedocs.io/en/latest/, and also see some other review issues in this repository for actual examples of JOSS reviews (in particular, reviewer's checklists). If that looks good and you agree to review for JOSS, that will be much appreciated. Then the review will start when we will have two reviewers.

I took a look at the criteria, and it should be doable, not until the end of the month, though. It still feels a bit strange that the reviewers are discussed in a GitHub chat open to the author.
