### 2024-05-25 #6327 @olexandr-konovalov
@YehorYudinIPP I see - I've found your details in our reviewers database, but it was not clear to me whether do you have any previous experience of reviewing for JOSS. Yes, that's a unique model (and I think open review perfectly aligns with the meaning of "O" in "JOSS"). It was described, for example, in https://doi.org/10.7717/peerj-cs.147 and https://doi.org/10.1109/MCSE.2018.03221930.

The review starts when I will find a 2nd reviewer, so it might not happen this month anyway.
