### 2024-06-04 a571962
adjusted check for symbolic math toolbox
