+++ src/misc/checkForPackage.m
@@ -1,0 +1,12 @@
+function isIncluded = checkForPackage(packageName)
+% checks if a particular package is installed 
+    packages = ver;
+    isIncluded = false;
+    for ip = packages
+        % check if name of element matches desired package
+        if strcmp(packages.Name, packageName)
+            isIncluded = true;
+            break;
+        end
+    end
+end
