### 2024-06-04 #6327 @olexandr-konovalov
Hello @fruzsinaagocs - we are looking for a reviewer for this paper. Would you be able to review this submission for JOSS, please? If not, could you possibly recommend someone whom we may approach regarding this?
