### 2024-06-19 #6902 @editorialbot
[REVIEW]: PathFinder: A Matlab/Octave package for oscillatory integration
**Submitting author:** @AndrewGibbs (<a href="http://orcid.org/0000-0002-2934-008X">Andrew Gibbs</a>)
**Repository:** https://github.com/AndrewGibbs/PathFinder
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
