### 2024-07-09 AndrewGibbs/PathFinder#16 @YehorYudinIPP
LAPACKE dependency
The Octave MEX functions compilation can crash if there is no LAPACKE installed.

It would be nice if the dependencies were listed somewhere in the installation instructions or if there was a script for requirements installation.
In this case, it could include `liblapack` `liblapack-dev` `libopenblas-base` `libopenblas-dev` `liblapacke-dev`
