### 2024-07-09 #6902 @YehorYudinIPP
Dear @olexandr-konovalov ,

in principle, I went through the checklist as well as explored and tried out the package, checked some of the references etc.

However, before checking in some of the list points, I would like to raise some questions. Should I simply openly create a repository issue for each of the points I have?

Best,
Yehor
