### 2024-07-12 AndrewGibbs/PathFinder#18 @YehorYudinIPP
Comments on the paper
Dear Andrew,

I have read your JOSS paper and have several comments and suggestions on it, mainly regarding the citing.

Firstly, it would be appreciated if the paper had DOI or other indicators for the cited papers.

Secondly, the paper would benefit from citing the Golub-Welsch algorithm implementation by Laurie and Gautschi, providing a more explicit description of this algorithm, as well as commenting on the Dijkstra algorithm.

Thirdly, it would be nice if the paper explicitly included the expressions of the solved integrals and necessary parameters like the integral domain and function calls to compute the integrals. This includes a concealed expression cited as the equation (32) for |𝐴32(𝑥1, 𝑥2)| from Hewett et al. (2019).

All the best,
Yehor
