### 2024-07-15 AndrewGibbs/PathFinder#19 @YehorYudinIPP
Repository suggestions
Dear Andrew,

Could you please tell me how one could formally contribute to your open-source software, report an issue or ask for support?
A short document with community software usage and development guidelines would be really helpful!

Also, it would be appreciated to have some more documentation (in comments) on function signature: arguments, their meaning and possible values.

All the best,
Yehor
