### 2024-07-23 #6902 @olexandr-konovalov
Many thanks, @YehorYudinIPP! Yes, we encourage you to submit issues to https://github.com/AndrewGibbs/PathFinder/issues - see https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. I've just noticed that you have already submitted some - if they have to be resolved before checking the remaining tickboxes, please add cross-references to link those issues to this reviews.
