### 2024-07-23 #6902 @YehorYudin
>Many thanks, @YehorYudinIPP! Yes, we encourage you to submit issues to https://github.com/AndrewGibbs/PathFinder/issues - see https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. I've just noticed that you have already submitted some - if they have to be resolved before checking the remaining tickboxes, please add cross-references to link those issues to this reviews.

Here is a summary of current suggestions from my side:

- [ ] Stating dependencies list or preparing an installation script: https://github.com/AndrewGibbs/PathFinder/issues/16
- [ ] Checking Octave plotting functionality and perfromance: https://github.com/AndrewGibbs/PathFinder/issues/16
- [ ] Checking the JOSS paper citation and adding information from the references: https://github.com/AndrewGibbs/PathFinder/issues/18
- [ ] Reporistory community guide and API documentation: https://github.com/AndrewGibbs/PathFinder/issues/19
