### 2024-08-20 #6902 @olexandr-konovalov
@fruzsinaagocs how are you getting on with this review? If you are still happy to review this paper for JOSS, could you please generate your checklist, as explained above, and start filling it in, to help us to see at which stage you're with the review? Please let me know if you have any questions about the process!
