### 2024-09-03 4828a34
added JOSS badge to README
