### 2024-10-04 #6902 @olexandr-konovalov
Thanks @fruzsinaagocs, for starting the review - how are you getting on with the rest of the checks?

In the meantime, I see that some issues by @YehorYudin are still open - apparently they are needed to complete @YehorYudin's review. @AndrewGibbs are you going to respond to them, or are you waiting to hear from @fruzsinaagocs first?
