### 2024-10-10 #6902 @AndrewGibbs
Oh dear! A colleague had an issue with his Mac installation but we decided it was a more general MEX-OSX issue. I will look into this next week and get back to you.
