### 2024-11-18 fd8d31b
removed mem dealloc in sd_path_ode_euler_extend_coarse_path.c
