+++ src/mex_compilation/sd_path_ode_euler_extend_coarse_path.c
@@ -105,9 +105,4 @@ void mexFunction(int nlhs, mxArray *plhs[],
         yn = 0;
     }
     convert_int_to_mxsca(&yn, plhs[2]);
-
-    free(p_log);
-    p_log = NULL;
-    free(h_log);
-    h_log = NULL;
 }
