### 2024-11-18 9499cf3
removed mem dealloc in sd_quad_ode_euler_correction.c
