### 2024-11-18 e0ab45a
added Mac-Octave special case to compileAll.m
