+++ compileAll.m
@@ -29,8 +29,14 @@ coversPath = [PFdir '/balls/'];
 
 if usingOctave()
   display("Compiling for Octave MEX");
-  mkoctfile('--mex','-llapacke', '-llapack', '-lblas', '-o', ...
+  if ismac
+        mkoctfile(...%'--mex','-llapacke', '-llapack', '-lblas', '-o', ...
+        '--mex','-o',[coversPath 'getSmallestSupsetBallMex'], [mexPath 'get_smallest_supset_ball.c'],...
+        '-I/opt/homebrew/opt/openblas/include','-L/opt/homebrew/opt/openblas/lib','-lopenblas');
+  else
+      mkoctfile('--mex','-llapacke', '-llapack', '-lblas', '-o', ...
       [coversPath 'getSmallestSupsetBallMex'], [mexPath 'get_smallest_supset_ball.c']);
+  end
 else
   display("Compiling for Matlab MEX");
   mex('-output',[coversPath 'getSmallestSupsetBallMex'], ...
