### 2024-11-18 ee1f257
Update CI.yml
Added Ubuntu and MacOS to test, and removed the Windows GCC installation step
