### 2024-11-18 0a81b52
Removed the Windows compiler switch in Test_all.m
