### 2024-11-18 1a2e0cb
continue-on-error in CI.yml
