+++ .github/workflows/CI.yml
@@ -6,7 +6,8 @@ jobs:
     strategy:
       matrix:
         os: [windows-latest, macos-latest, ubuntu-latest]
-    runs-on: ${{ matrix.os }}   
+    runs-on: ${{ matrix.os }}  
+    continue-on-error: true
     steps:
       - name: Check out repository
         uses: actions/checkout@v3
