### 2024-11-18 bea8d07
Re-added GCC for Win to CI.yml
