### 2024-11-18 904c2e5
Re-added PC compiler switch in Test_all.m
