### 2024-11-18 d1029cd
Update CI.yml
adding ubuntu to CI
