### 2024-11-18 1ec806f
Create ubuntusolo.yml
