### 2024-11-18 3f952be
Update and rename CI.yml to windows.yml
Moving to separate YML files for each test, the matrix OS approach seems a bit dodgy
