### 2024-11-18 8c141a8
Update and rename ubuntusolo.yml to ubuntu.yml
