+++ .github/workflows/ubuntu.yml
@@ -2,7 +2,7 @@ name: Ubuntu-solo
 on: [push]
 jobs:
   my-job:
-    name: Run MATLAB Tests and Generate Artifacts
+    name: Run MATLAB Tests and Generate Artifacts on Ubuntu
     runs-on: ubuntu-latest       
     steps:
       - name: Check out repository
