+++ .github/workflows/Mac.yml
@@ -1,0 +1,20 @@
+name: Mac-solo
+on: [push]
+jobs:
+  my-job:
+    name: Run MATLAB Tests and Generate Artifacts on MacOS
+    runs-on: macos-latest       
+    steps:
+      - name: Check out repository
+        uses: actions/checkout@v3
+      - name: Set up MATLAB
+        uses: matlab-actions/setup-matlab@v1
+      - name: Run tests and generate artifacts
+        uses: matlab-actions/run-tests@v1
+        with:
+          test-results-junit: test-results/results.xml
+          code-coverage-cobertura: code-coverage/coverage.xml
+      - name: Upload coverage reports to Codecov
+        uses: codecov/codecov-action@v3
+        env:
+          CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
