+++ .github/workflows/Mac.yml
@@ -3,7 +3,7 @@ on: [push]
 jobs:
   my-job:
     name: Run MATLAB Tests and Generate Artifacts on MacOS
-    runs-on: macos-latest       
+    runs-on: macos-14    
     steps:
       - name: Check out repository
         uses: actions/checkout@v3
