+++ .github/workflows/Mac.yml
@@ -8,7 +8,7 @@ jobs:
       - name: Check out repository
         uses: actions/checkout@v3
       - name: Set up MATLAB
-        uses: matlab-actions/setup-matlab@v1
+        uses: matlab-actions/setup-matlab@v2
       - name: Run tests and generate artifacts
         uses: matlab-actions/run-tests@v1
         with:
