### 2024-11-18 0581f7d
Delete .github/workflows/Mac.yml
