### 2024-11-18 be54b58
Update README.md for separate tests
