+++ .github/workflows/windows.yml
@@ -1,4 +1,4 @@
-name: CI
+name: Windows
 on: [push]
 jobs:
   my-job:
