### 2024-11-18 b0e42be
Update ubuntu.yml
