+++ .github/workflows/ubuntu.yml
@@ -1,4 +1,4 @@
-name: Ubuntu-solo
+name: Ubuntu
 on: [push]
 jobs:
   my-job:
