### 2024-11-18 6c635e0
Update draft-pdf.yml
Was bug in draft JOSS paper creation, see here for details:
https://github.blog/changelog/2024-02-13-deprecation-notice-v1-and-v2-of-the-artifact-actions/
