+++ .github/workflows/draft-pdf.yml
@@ -13,7 +13,7 @@ jobs:
           journal: joss
           paper-path: documents/joss_paper/paper.md
       - name: Upload
-        uses: actions/upload-artifact@v1
+        uses: actions/upload-artifact@v3
         with:
           name: paper
           # This is the output path where Pandoc will write the compiled
