### 2024-11-18 924b91c
Update draft-pdf.yml
... actually v4 of the artifact download is suggested
