### 2024-11-20 #6902 @fruzsinaagocs
The installation issues have now been resolved, so I'm continuing my review of the code.
The paper already contains comparisons with other software performing oscillatory integrals, but the ones below might also be worth mentioning (without detailed comparisons):
- [OscillatoryIntegralsODE.jl](https://github.com/xzackli/OscillatoryIntegralsODE.jl),
- [Picard-Lefschetz Path Integrals](https://github.com/P-LPI/P-LPI.github.io) (with further description [here](https://p-lpi.github.io)).
