### 2024-12-09 #6902 @olexandr-konovalov
Thank you @fruzsinaagocs ! From your checklist it seems like lots of things have been checked. Do you think you will be able to finish the review soon, and let the author now what has to be done to check the remaining boxes, if any will need to be addressed after your review?
