### 2024-12-27 #6902 @fruzsinaagocs
I have the following items left on my checklist that need addressing by the authors:
1. Community guidelines: please specify how users can contribute to the project, report issues, and ask for support.
2. Functionality documentation: it would be nice to have API-level documentation, if possible.
3. References: ideally, all articles should have a DOI listed with them.

1 and 2 above have been pointed out by @YehorYudinIPP in [this issue](https://github.com/AndrewGibbs/PathFinder/issues/19#issue-2408798303), and 3 in [this one](https://github.com/AndrewGibbs/PathFinder/issues/18#issue-2405958186).
