### 2025-03-23 #6902 @AndrewGibbs
Hi Daniel,

Please accept my apologies for the delay. Due to a change in my circumstances, I only have time to work on this during the weekends now. The reviewers' suggestions are great, but it will take me at least a month before I can find time to implement them.

Can I ask, what is the process for responding to the reviewers' suggestions, especially in the small number of cases where their suggestions are not possible? Shall I just reply to their comments on here, or on the repository? Or should I write a full summary on here?

Best wishes,

Andrew
