### 2025-03-30 #6902 @olexandr-konovalov
Thank you for the update @AndrewGibbs! As the editorial guide says at https://joss.readthedocs.io/en/latest/editing.html#review

> Comments in the REVIEW issue should be kept brief, as much as possible, with more lengthy suggestions or requests posted as separate issues, directly in the submission repository. A link-back to those issues in the REVIEW is helpful.

Having most of the discussions in the issues in your repository will help to keep them close to your code, while links from the review issue will help to check your responses. This also makes sense in the specific case where implementing the suggestion is not possible.

I hope this answers the question - please point out to the specific suggestions and/or issues in case of any questions.
