### 2025-05-06 #6902 @AndrewGibbs
Dear @YehorYudinIPP  and @fruzsinaagocs,

I've made many of the amendments you suggested, thank you for all of these. Those which I have been unable to action are discussed on the PathFinder [Issues page.](https://github.com/AndrewGibbs/PathFinder/issues)  The main difference since the first submission: documentation for the code now takes the form of a Jupyter book, rather than being squeezed into a README; this felt like the best way to make all of the necessary improvements.

Fruzsina, you made an [additional suggestion](https://github.com/openjournals/joss-reviews/issues/6902#issuecomment-2487041550) about related software. Thanks for this. I have added a new subsection for each of these packages at the end of the paper.

Thank you both for your patience. I look forward to hearing your response.
