### 2025-05-06 #6902 @olexandr-konovalov
Nice to hear about further progress @AndrewGibbs! You need to give commands to the editorial bot as separate comments to make them work.
