### 2025-06-19 #6902 @fruzsinaagocs
Hi @olexandr-konovalov, thank you for the reminder, I'll check the updates soon.
