### 2025-06-30 #6902 @YehorYudinIPP
I am happy with the state of the paper and documentation, and I give my green light!
