### 2025-08-05 #6902 @olexandr-konovalov
Nice, thanks @fruzsinaagocs! Next, @AndrewGibbs please add missing DOIs and let me know when done.
