### 2025-08-11 #6902 @fruzsinaagocs
> Thank you [@fruzsinaagocs](https://github.com/fruzsinaagocs) . May I ask, how did you find the DOIs for these articles? I tried but failed!

No problem! I just searched for "<article title> doi" which usually landed me on the publisher's page. The DOI was then typically  hidden somewhere on that page.
