### 2025-09-05 #6902 @AndrewGibbs
Dear @olexandr-konovalov ,

I have archived on Zenodo. The DOI is: **10.5281/zenodo.17058699**

One question: The license is MIT on GitHub and Zenodo. I have noticed on the JOSS paper it is a CC license. Is this the license for the paper its self, as opposed to the software? Or should this match too?

Best wishes,

Andrew
