### 2025-09-05 #6902 @danielskatz
To jump in again, the CC license for the paper is the license for just the paper itself.
