### 2025-10-26 #6902 @danielskatz
@AndrewGibbs - please check this and make sure it seems ok to you
