### 2025-10-26 #6902 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
