### 2025-10-29 #6902 @AndrewGibbs
On a final check I noticed a typo in the title of Figure 1: 'catast**r**ophe' was misspelled. I have fixed this, and the code used to create it in `examples/custCatastrophe.m`. All happy to proceed now, thank you for your patience!
