### 2025-10-30 #6902 @danielskatz
Congratulations to @AndrewGibbs (Andrew Gibbs) on your JOSS publication!!

JOSS completely depends on volunteers and we couldn't do this without your work.
