### 2025-10-30 #6902 @danielskatz
It works fine for me, and did when I closed this issue. I think you're getting some DNS caching problem and it will work for you in a little while
